+++ src/main/java/com/github/tommyettinger/ds/IntFloatMap.java
@@ -141,7 +141,7 @@ public class IntFloatMap implements Iterable<IntFloatMap.Entry> {
 		int tableSize = tableSize(initialCapacity, loadFactor);
 		threshold = (int)(tableSize * loadFactor);
 		mask = tableSize - 1;
-		shift = BitConversion.countLeadingZeros(mask);
+		shift = BitConversion.countLeadingZeros((long)mask);
 
 		keyTable = new int[tableSize];
 		valueTable = new float[tableSize];
@@ -621,7 +621,7 @@ public class IntFloatMap implements Iterable<IntFloatMap.Entry> {
 		int oldCapacity = keyTable.length;
 		threshold = (int)(newSize * loadFactor);
 		mask = newSize - 1;
-		shift = BitConversion.countLeadingZeros(mask);
+		shift = BitConversion.countLeadingZeros((long)mask);
 
 		hashMultiplier = Utilities.GOOD_MULTIPLIERS[(int)(hashMultiplier >>> 48 + shift) & 511];
 		int[] oldKeyTable = keyTable;
